+++ tests/internals/custom_hpa_name/custom_hpa_name_test.go
@@ -46,11 +46,6 @@ spec:
         image: registry.k8s.io/hpa-example
         ports:
         - containerPort: 80
-        resources:
-          limits:
-            cpu: 500m
-          requests:
-            cpu: 200m
         imagePullPolicy: IfNotPresent
 `
 
@@ -68,10 +63,12 @@ spec:
   maxReplicaCount: 1
   cooldownPeriod: 10
   triggers:
-    - type: cpu
-      metadata:
-        type: Utilization
-        value: "50"
+  - type: metrics-api
+    metadata:
+      targetValue: "2"
+      url: "invalid-invalid"
+      valueLocation: 'value'
+      method: "query"
 `
 
 	scaledObjectTemplateWithCustomName = `
@@ -90,10 +87,12 @@ spec:
     horizontalPodAutoscalerConfig:
       name: {{.CustomHpaName}}
   triggers:
-      - type: cpu
-        metadata:
-          type: Utilization
-          value: "50"
+  - type: metrics-api
+    metadata:
+      targetValue: "2"
+      url: "invalid-invalid"
+      valueLocation: 'value'
+      method: "query"
 `
 )