+++ tests/scalers/azure/azure_event_hub_aad_wi/azure_event_hub_aad_wi_test.go
@@ -97,7 +97,6 @@ spec:
       containers:
         - name: {{.DeploymentName}}
           image: ghcr.io/kedacore/tests-azure-eventhub-golang
-          resources:
           env:
             - name: EVENTHUB_CONNECTION_STRING
               valueFrom: